Populate job keywords during data refresh
Move keyword population out of parsing process because it's not logically part of parsing the data feed.
