+++ commercial/app/model/commercial/jobs/Job.scala
@@ -2,12 +2,7 @@ package model.commercial.jobs
 
 import model.commercial.{Keyword, Ad, Segment}
 import model.commercial.Utils._
-import conf.ContentApi
-import scala.concurrent.Await
-import scala.concurrent.duration._
-import common.{Logging, ExecutionContexts}
-import scala.util.{Success, Failure, Try}
-import scala.collection.mutable
+import common.ExecutionContexts
 
 case class Job(id: Int,
                title: String,
@@ -26,9 +21,9 @@ case class Job(id: Int,
   }
 }
 
-object Job extends ExecutionContexts with Logging {
+object Job extends ExecutionContexts {
 
-  private val sectorIdIndustryMap = Map[Int, String](
+  val sectorIdIndustryMap = Map[Int, String](
     (101, "Arts & heritage"),
     (111, "Charities"),
     (124, "Construction"),
@@ -55,42 +50,4 @@ object Job extends ExecutionContexts with Logging {
     (343, "Skilled Trade"),
     (350, "Social Enterprise")
   )
-
-  private def lookUpCorrespondingKeywords(term: String): Seq[Keyword] = {
-    val q = term replace("&", "") replace(",", "")
-    Try(Await.result(
-      ContentApi.tags.stringParam("type", "keyword").stringParam("q", q).pageSize(50).response map {
-        _.results map (tag => Keyword(tag.id, tag.webTitle))
-      }, 2.seconds)) match {
-      case Success(result) => result
-      case Failure(e) =>
-        log.warn(s"Looking up $term failed: $e")
-        Nil
-    }
-  }
-
-  private def sectorIdKeywords(sectorIds: Seq[Int]): Map[Int, Seq[Keyword]] = {
-    val industryKeywords = mutable.Map[String, Seq[Keyword]]()
-    sectorIds.foldLeft(Map[Int, Seq[Keyword]]()) {
-      (acc, sectorId) => {
-        val keywords = sectorIdIndustryMap.get(sectorId).map {
-          industry =>
-            industryKeywords.getOrElseUpdate(industry, lookUpCorrespondingKeywords(industry))
-        }.getOrElse(Nil)
-        acc + (sectorId -> keywords)
-      }
-    }
-  }
-
-  def populateKeywords(jobs: Seq[Job]): Seq[Job] = {
-    val allSectorIds = jobs.flatMap(_.sectorIds)
-    val keywords = sectorIdKeywords(allSectorIds)
-    jobs.map {
-      job =>
-        val jobKeywords = job.sectorIds.foldLeft(Seq[Keyword]()) {
-          (acc, sectorId) => (acc ++ keywords.get(sectorId).getOrElse(Nil)).distinct
-        }
-        job.copy(keywords = jobKeywords)
-    }
-  }
 }
